requirements.txt
alpaca==0.8.2
pytorch==2.2.2+cu118
numpy==1.24.3
matplotlib==3.7.1
pandas==1.5.3
toml==0.10.2
statsmodels==0.14.0